+++ legacy-data-access-services/src/test/resources/hibernate.cfg.xml
@@ -60,7 +60,7 @@
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.ClientServiceProvider"/>
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.OtherClientName" />
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.TribalMembershipVerification" />
-    <mapping class="gov.ca.cwds.cms.data.access.service.impl.dao.BaseClientImpl" />
+    <mapping class="gov.ca.cwds.data.persistence.cms.BaseClientImpl" />
 
 
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ActiveServiceComponentType"/>
